second attempt; check_param flag.  Compiles, works as well as before... but doesn't seem to solve the mysteries I expected it to.

+++ dynchk/libs/libc/kill.c
@@ -7,19 +7,21 @@
 #undef kill
 static int(*funcptr) (pid_t , int ) = 0;
 
+extern int __lsb_check_params;
 int kill (pid_t arg0 , int arg1 )
 {
+	int reset_flag = __lsb_check_params;
+	int ret_value  ;
 	if(!funcptr)
 		funcptr = dlsym(RTLD_NEXT, "kill");
-	validate_NULL_TYPETYPE(  arg0, "kill - arg0");
-	validate_NULL_TYPETYPE(  arg1, "kill - arg1");
-	return funcptr(arg0, arg1);
-}
-
-int __lsb_kill (pid_t arg0 , int arg1 )
-{
-	if(!funcptr)
-		funcptr = dlsym(RTLD_NEXT, "kill");
-	return funcptr(arg0, arg1);
+	if(__lsb_check_params)
+	{
+		__lsb_check_params=0;
+		validate_NULL_TYPETYPE(  arg0, "kill - arg0");
+		validate_NULL_TYPETYPE(  arg1, "kill - arg1");
+	}
+	ret_value = funcptr(arg0, arg1);
+	__lsb_check_params = reset_flag;
+	return ret_value;
 }
